src/python/requirements.txt
# MATLAB Engine API for Python - Requirements
# For Issue #1 Implementation

# Core MATLAB Engine API
matlabengine>=25.1.2

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development Tools
black>=23.0.0  # Code formatting
pylint>=2.17.0  # Code linting
mypy>=1.4.0  # Type checking

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Optional: Performance Monitoring
psutil>=5.9.0  # System resource monitoring
memory-profiler>=0.61.0  # Memory usage profiling